some changes

+++ backend/src/main/java/cs/vsu/raspopov/carkit/repository/AutoRepairShopRepo.java
@@ -3,5 +3,5 @@ package cs.vsu.raspopov.carkit.repository;
 import cs.vsu.raspopov.carkit.entity.AutoRepairShop;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AutoRepairShopRepo extends CrudRepository<AutoRepairShop, Integer> {
+public interface AutoRepairShopRepo extends CrudRepository<AutoRepairShop, Long> {
 }
