+++ backend/src/main/java/cs/vsu/raspopov/carkit/repository/DetailRepo.java
@@ -3,6 +3,6 @@ package cs.vsu.raspopov.carkit.repository;
 import cs.vsu.raspopov.carkit.entity.Detail;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DetailRepo extends CrudRepository<Detail, Integer> {
+public interface DetailRepo extends CrudRepository<Detail, Long> {
 
 }
