+++ backend/src/main/java/cs/vsu/raspopov/carkit/repository/ModificationRepo.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.Optional;
 
-public interface ModificationRepo extends CrudRepository<Modification, Integer> {
+public interface ModificationRepo extends CrudRepository<Modification, Long> {
 
     Optional<Modification> findByNameAndEngineModelAndYearFromAndYearTo(
             String name, String engineModel, Date yearFrom, Date yearTo);
