+++ backend/src/main/java/cs/vsu/raspopov/carkit/repository/ScheduleRepo.java
@@ -3,5 +3,5 @@ package cs.vsu.raspopov.carkit.repository;
 import cs.vsu.raspopov.carkit.entity.Schedule;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ScheduleRepo extends CrudRepository<Schedule, Integer> {
+public interface ScheduleRepo extends CrudRepository<Schedule, Long> {
 }
